Implement localeCompare for string sorting, and dedicated comparisongs for boolean and date fields
